+++ .github/workflows/python-app.yml
@@ -31,9 +31,9 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --exclude=./subtitles/subtitles_pb2.py,./subtitles/subtitles_pb2_grpc.py,./mock_receiver/subtitles_pb2.py,./mock_receiver/subtitles_pb2_grpc.py --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 . --exclude=./voice/subtitles_pb2.py,./voice/subtitles_pb2_grpc.py,./voice/audio_pb2.py,./voice/audio_pb2_grpc.py,./mock_receiver/subtitles_pb2.py,./mock_receiver/subtitles_pb2_grpc.py --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --exclude=./subtitles/subtitles_pb2.py,./subtitles/subtitles_pb2_grpc.py,./mock_receiver/subtitles_pb2.py,./mock_receiver/subtitles_pb2_grpc.py --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 . --exclude=./voice/subtitles_pb2.py,./voice/subtitles_pb2_grpc.py,./voice/audio_pb2.py,./voice/audio_pb2_grpc.py,./mock_receiver/subtitles_pb2.py,./mock_receiver/subtitles_pb2_grpc.py --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     # - name: Test with pytest
     #  run: |
     #    pytest
