+++ voice/properties.py
